requirements.txt
Flask==3.1.3
Werkzeug==3.1.3
pywebview==5.3.1
